# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/auth/token": {
            "post": {
                "description": "Exchanges a valid API key for a signed JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Issue Token",
                "parameters": [
                    {
                        "description": "Issue token request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.IssueTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.IssueTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/executions": {
            "get": {
                "description": "Returns executions with optional filters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "List Executions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set to summary for summary view",
                        "name": "view",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by DAG ID",
                        "name": "dag_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by execution status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination limit (for dag_id filter)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (for dag_id filter)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execution list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}": {
            "get": {
                "description": "Returns execution details by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "Get Execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execution",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/executions/{id}/nodes": {
            "get": {
                "description": "Returns node-level results for one execution.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "Get Execution Nodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Execution node results",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/executions/{id}/resume": {
            "post": {
                "description": "Resumes a suspended execution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "Resume Execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional human intervention context",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.ResumeExecutionRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/dto.RunExecutionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/run": {
            "post": {
                "description": "Accepts a full DAG configuration and executes it immediately.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "Run Inline DAG",
                "parameters": [
                    {
                        "description": "DAG configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/dto.RunExecutionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.IssueTokenRequest": {
            "type": "object",
            "required": [
                "api_key"
            ],
            "properties": {
                "api_key": {
                    "type": "string"
                }
            }
        },
        "dto.IssueTokenResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResumeExecutionRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "actor": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "dto.RunExecutionResponse": {
            "type": "object",
            "properties": {
                "execution_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ExecutionStatus"
                }
            }
        },
...
        "models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "timeout",
                "suspended"
            ],
            "x-enum-comments": {
                "StatusSuspended": "waiting for human approval"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "waiting for human approval"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusRunning",
                "StatusCompleted",
                "StatusFailed",
                "StatusTimeout",
                "StatusSuspended"
            ]
        },
...
    }
}